put logic in component
