+++ demo/index.html
@@ -256,9 +256,9 @@
     <section>
       <h2>Audio visualiser</h2>
       <h3>Local audio</h3>
-      <canvas data-selector="local-audio" width="1024" height="128a"></canvas>
+      <canvas data-selector="localAudio" width="1024" height="128a"></canvas>
       <h3>Remote audio</h3>
-      <canvas data-selector="remote-audio" width="1024" height="128a"></canvas>
+      <canvas data-selector="remoteAudio" width="1024" height="128a"></canvas>
     </section>
   </template>
